+++ src/posix/cocoa/i_main.mm
@@ -31,10 +31,9 @@
  **
  */
 
-#include <sys/sysctl.h>
+#include "i_common.h"
 
-#import <AppKit/NSMenu.h>
-#import <AppKit/NSEvent.h>
+#include <sys/sysctl.h>
 
 // Avoid collision between DObject class and Objective-C
 #define Class ObjectClass
@@ -48,9 +47,6 @@
 #include "m_argv.h"
 #include "s_sound.h"
 #include "version.h"
-
-#include "i_common.h"
-#include "i_osversion.h"
 
 #undef Class
 
@@ -492,27 +488,6 @@ void CreateMenu()
 	[NSApp setMainMenu:menuBar];
 }
 
-DarwinVersion GetDarwinVersion()
-{
-	DarwinVersion result = {};
-
-	int mib[2] = { CTL_KERN, KERN_OSRELEASE };
-	size_t size = 0;
-
-	if (0 == sysctl(mib, 2, NULL, &size, NULL, 0))
-	{
-		char* version = static_cast<char*>(alloca(size));
-
-		if (0 == sysctl(mib, 2, version, &size, NULL, 0))
-		{
-			sscanf(version, "%hu.%hu.%hu",
-				&result.major, &result.minor, &result.bugfix);
-		}
-	}
-
-	return result;
-}
-
 void ReleaseApplicationController()
 {
 	if (NULL != appCtrl)
@@ -526,9 +501,6 @@ void ReleaseApplicationController()
 }
 
 } // unnamed namespace
-
-
-const DarwinVersion darwinVersion = GetDarwinVersion();
 
 
 int main(int argc, char** argv)
